more

+++ src/python/py/models/builder.py
@@ -653,7 +653,7 @@ class Model:
         # NOTE: Use ast.literal_eval instead of eval as eval allows arbitrary code execution
         scalar_or_array = ast.literal_eval(path[-1])
         assert dims in {"0D", "1D"}, f"Unexpected dimension {dims} in constant name"
-        if dims == "1D" and isinstance(scalar_or_array, (int, float)):
+        if dims == "1D" and isinstance(scalar_or_array, (float, int)):
             # Convert to 1D tensor
             scalar_or_array = [scalar_or_array]
         tensor = ir.tensor(scalar_or_array, dtype=onnx_dtype, name=name)
